Refactor  test infra
