+++ packages/flutter_adaptive_scaffold/test/test_breakpoints.dart
@@ -1,0 +1,39 @@
+import 'package:flutter/widgets.dart';
+import 'package:flutter_adaptive_scaffold/src/breakpoints.dart';
+
+class TestBreakpoint0 extends Breakpoint {
+  @override
+  bool isActive(BuildContext context) {
+    return MediaQuery.of(context).size.width >= 0 &&
+        MediaQuery.of(context).size.width < 800;
+  }
+}
+
+class TestBreakpoint400 extends Breakpoint {
+  @override
+  bool isActive(BuildContext context) {
+    return MediaQuery.of(context).size.width > 400;
+  }
+}
+
+class TestBreakpoint800 extends Breakpoint {
+  @override
+  bool isActive(BuildContext context) {
+    return MediaQuery.of(context).size.width >= 800 &&
+        MediaQuery.of(context).size.width < 1000;
+  }
+}
+
+class TestBreakpoint1000 extends Breakpoint {
+  @override
+  bool isActive(BuildContext context) {
+    return MediaQuery.of(context).size.width >= 1000;
+  }
+}
+
+class NeverOnBreakpoint extends Breakpoint {
+  @override
+  bool isActive(BuildContext context) {
+    return false;
+  }
+}
